Modify iOS for bridge perf
