+++ examples/bridge_echo/iOS/BridgePerf.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
