+++ examples/bridge_echo/iOS/BridgePerf/Info.plist
